+++ admin/src/main/java/com/liulin/web/controller/moduler/data/ResidentController.java
@@ -140,8 +140,8 @@ public class ResidentController extends BaseController
         String attachmentIds = resident.getAttachmentIds();
         if(StringUtils.isNotEmpty(attachmentIds)){
             String[] attachmentIdStrArray = attachmentIds.split(",");
-            int [] attachmentIdArray =
-                    Arrays.asList(attachmentIdStrArray).stream().mapToInt(Integer::parseInt).toArray();
+            long [] attachmentIdArray =
+                    Arrays.asList(attachmentIdStrArray).stream().mapToLong(Long::parseLong).toArray();
             List<Attachment> attachments = attachmentService.selectAttachmentByIds(attachmentIdArray);
             if(CollectionUtils.isNotEmpty(attachments)){
                 mmap.put("attachments",attachments);
