+++ admin/src/main/java/com/liulin/web/controller/moduler/data/ServizeController.java
@@ -1,32 +1,26 @@
 package com.liulin.web.controller.moduler.data;
 
-import java.util.Arrays;
-import java.util.List;
-
+import com.liulin.common.annotation.Log;
+import com.liulin.common.core.controller.BaseController;
+import com.liulin.common.core.domain.AjaxResult;
+import com.liulin.common.core.page.TableDataInfo;
+import com.liulin.common.enums.BusinessType;
 import com.liulin.common.utils.ShiroUtils;
 import com.liulin.common.utils.StringUtils;
-import com.liulin.system.domain.Asset;
+import com.liulin.common.utils.poi.ExcelUtil;
 import com.liulin.system.domain.Attachment;
-import com.liulin.system.domain.Resident;
+import com.liulin.system.domain.Servize;
 import com.liulin.system.service.IAttachmentService;
+import com.liulin.system.service.IServizeService;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import com.liulin.common.annotation.Log;
-import com.liulin.common.enums.BusinessType;
-import com.liulin.system.domain.Servize;
-import com.liulin.system.service.IServizeService;
-import com.liulin.common.core.controller.BaseController;
-import com.liulin.common.core.domain.AjaxResult;
-import com.liulin.common.utils.poi.ExcelUtil;
-import com.liulin.common.core.page.TableDataInfo;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * serviceController
@@ -115,8 +109,8 @@ public class ServizeController extends BaseController
         String attachmentIds = servize.getAttachmentIds();
         if(StringUtils.isNotEmpty(attachmentIds)){
             String[] attachmentIdStrArray = attachmentIds.split(",");
-            int [] attachmentIdArray =
-                    Arrays.asList(attachmentIdStrArray).stream().mapToInt(Integer::parseInt).toArray();
+            long [] attachmentIdArray =
+                    Arrays.asList(attachmentIdStrArray).stream().mapToLong(Long::parseLong).toArray();
             List<Attachment> attachments = attachmentService.selectAttachmentByIds(attachmentIdArray);
             if(CollectionUtils.isNotEmpty(attachments)){
                 mmap.put("attachments",attachments);
