+++ admin/src/main/java/com/liulin/web/controller/moduler/data/SupplierController.java
@@ -143,8 +143,8 @@ public class SupplierController extends BaseController
         String attachmentIds = supplier.getAttachmentIds();
         if(StringUtils.isNotEmpty(attachmentIds)){
             String[] attachmentIdStrArray = attachmentIds.split(",");
-            int [] attachmentIdArray =
-                    Arrays.asList(attachmentIdStrArray).stream().mapToInt(Integer::parseInt).toArray();
+           long [] attachmentIdArray =
+                    Arrays.asList(attachmentIdStrArray).stream().mapToLong(Long::parseLong).toArray();
             List<Attachment> attachments = attachmentService.selectAttachmentByIds(attachmentIdArray);
             if(CollectionUtils.isNotEmpty(attachments)){
                 mmap.put("attachments",attachments);
