+++ admin/src/main/java/com/liulin/web/controller/system/SysDeptController.java
@@ -105,8 +105,8 @@ public class SysDeptController extends BaseController
         String attachmentIds = dept.getAttachmentIds();
         if(StringUtils.isNotEmpty(attachmentIds)){
             String[] attachmentIdStrArray = attachmentIds.split(",");
-            int [] attachmentIdArray =
-                    Arrays.asList(attachmentIdStrArray).stream().mapToInt(Integer::parseInt).toArray();
+           long [] attachmentIdArray =
+                    Arrays.asList(attachmentIdStrArray).stream().mapToLong(Long::parseLong).toArray();
             List<Attachment> attachments = attachmentService.selectAttachmentByIds(attachmentIdArray);
             if(CollectionUtils.isNotEmpty(attachments)){
                 mmap.put("attachments",attachments);
